# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  app_internal_interfaces_http_dto_request.ProductQuestionAnswerRequest:
    properties:
      answer:
        type: string
    required:
    - answer
    type: object
  app_internal_interfaces_http_dto_request.ProductQuestionAskRequest:
    properties:
      asker_email:
        maxLength: 255
        type: string
      asker_name:
        maxLength: 255
        type: string
      question:
        type: string
    required:
    - asker_name
    - question
    type: object
...
  app_internal_interfaces_http_dto_response.AdminProductQuestionListItemResponse:
    properties:
      answer:
        type: string
      answered_at:
        type: string
      answered_by:
        type: string
      asker_email:
        type: string
      asker_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      question:
        type: string
      status:
        type: string
    type: object
  app_internal_interfaces_http_dto_response.AdminProductQuestionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/app_internal_interfaces_http_dto_response.AdminProductQuestionListItemResponse'
        type: array
      meta:
        $ref: '#/definitions/app_pkg_pagination.Meta'
    type: object
...
  app_internal_interfaces_http_dto_response.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: request validation failed
        type: string
    type: object
  app_internal_interfaces_http_dto_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorBody'
      path:
        example: /api/v1/auth/login
        type: string
      statusCode:
        example: 400
        type: integer
    type: object
...
  app_internal_interfaces_http_dto_response.ProductQuestionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/app_internal_interfaces_http_dto_response.ProductQuestionResponse'
        type: array
      meta:
        $ref: '#/definitions/app_pkg_pagination.Meta'
    type: object
  app_internal_interfaces_http_dto_response.ProductQuestionResponse:
    properties:
      answer:
        type: string
      answered_at:
        type: string
      answered_by:
        type: string
      asker_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      question:
        type: string
      status:
        type: string
    type: object
...
  app_pkg_pagination.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@shop.com
    name: API Support
  description: REST API for the eCommerce admin panel and public storefront. Supports
    JWT authentication with role-based access (admin, customer). Store endpoints under
    /api/v1/store are public; account routes require a customer token.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: eCommerce API
  version: "1.0"
paths:
...
  /api/v1/admin/questions:
    get:
      description: Get a paginated list of product questions.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Product ID filter
        in: query
        name: product_id
        type: string
      - description: Status filter (open/answered)
        in: query
        name: status
        type: string
      - description: Search query by question content or asker name
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.AdminProductQuestionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List questions (Admin)
      tags:
      - admin-questions
  /api/v1/admin/questions/{id}:
    delete:
      description: Delete a product question.
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete question (Admin)
      tags:
      - admin-questions
  /api/v1/admin/questions/{id}/answer:
    post:
      consumes:
      - application/json
      description: Answer a product question.
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Answer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.ProductQuestionAnswerRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Answer question (Admin)
      tags:
      - admin-questions
...
  /api/v1/store/products/{productId}/questions:
    get:
      description: Get a paginated list of answered questions for a product.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ProductQuestionListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: List questions
      tags:
      - store-questions
    post:
      consumes:
      - application/json
      description: Submit a question about a product.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: Question data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.ProductQuestionAskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ProductQuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: Ask a question
      tags:
      - store-questions
...
securityDefinitions:
  BearerAuth:
    description: 'Enter "Bearer" followed by a space and the JWT access token. Example:
      Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"